+++ src/main/java/fun/gen/DoubleGen.java
@@ -41,25 +41,25 @@ public final class DoubleGen implements Gen<Double> {
                                      final double max) {
         if (max < min) throw new IllegalArgumentException("max < min");
         List<Pair<Integer, Gen<? extends Double>>> gens = new ArrayList<>();
-        if (max > Integer.MAX_VALUE && min < Integer.MAX_VALUE)
+        if (max >= Integer.MAX_VALUE && min <= Integer.MAX_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((double) Integer.MAX_VALUE)));
-        if (max > Integer.MIN_VALUE && min < Integer.MIN_VALUE)
+        if (max >= Integer.MIN_VALUE && min <= Integer.MIN_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((double) Integer.MIN_VALUE)));
-        if (max > Short.MAX_VALUE && min < Short.MAX_VALUE)
+        if (max >= Short.MAX_VALUE && min <= Short.MAX_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((double) Short.MAX_VALUE)));
-        if (max > Short.MIN_VALUE && min < Short.MIN_VALUE)
+        if (max >= Short.MIN_VALUE && min <= Short.MIN_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((double) Short.MIN_VALUE)));
-        if (max > Byte.MAX_VALUE && min < Byte.MAX_VALUE)
+        if (max >= Byte.MAX_VALUE && min <= Byte.MAX_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((double) Byte.MAX_VALUE)));
-        if (max > Byte.MIN_VALUE && min < Byte.MIN_VALUE)
+        if (max >= Byte.MIN_VALUE && min <= Byte.MIN_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((double) Byte.MIN_VALUE)));
-        if (max > 0 && min < 0)
+        if (max >= 0 && min <= 0)
             gens.add(Pair.of(1,
                              Gen.cons(0.0)));