+++ src/main/java/fun/gen/IntGen.java
@@ -58,19 +58,19 @@ public final class IntGen implements Gen<Integer> {
         if (max < min) throw new IllegalArgumentException("max < min");
         List<Pair<Integer, Gen<? extends Integer>>> gens = new ArrayList<>();
 
-        if (max > Short.MAX_VALUE && min < Short.MAX_VALUE)
+        if (max >= Short.MAX_VALUE && min <= Short.MAX_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((int) Short.MAX_VALUE)));
-        if (max > Short.MIN_VALUE && min < Short.MIN_VALUE)
+        if (max >= Short.MIN_VALUE && min <= Short.MIN_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((int) Short.MIN_VALUE)));
-        if (max > Byte.MAX_VALUE && min < Byte.MAX_VALUE)
+        if (max >= Byte.MAX_VALUE && min <= Byte.MAX_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((int) Byte.MAX_VALUE)));
-        if (max > Byte.MIN_VALUE && min < Byte.MIN_VALUE)
+        if (max >= Byte.MIN_VALUE && min <= Byte.MIN_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((int) Byte.MIN_VALUE)));
-        if (max > 0 && min < 0)
+        if (max >= 0 && min <= 0)
             gens.add(Pair.of(1,
                              Gen.cons(0)));