+++ src/main/java/fun/gen/LongGen.java
@@ -42,25 +42,26 @@ public final class LongGen implements Gen<Long> {
         if (max < min) throw new IllegalArgumentException("max < min");
 
         List<Pair<Integer, Gen<? extends Long>>> gens = new ArrayList<>();
-        if (max > Integer.MAX_VALUE && min < Integer.MAX_VALUE)
+        if (max >= Integer.MAX_VALUE && min <= Integer.MAX_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((long) Integer.MAX_VALUE)));
-        if (max > Integer.MIN_VALUE && min < Integer.MIN_VALUE)
+
+        if (max >= Integer.MIN_VALUE && min <= Integer.MIN_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((long) Integer.MIN_VALUE)));
-        if (max > Short.MAX_VALUE && min < Short.MAX_VALUE)
+        if (max >= Short.MAX_VALUE && min <= Short.MAX_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((long) Short.MAX_VALUE)));
-        if (max > Short.MIN_VALUE && min < Short.MIN_VALUE)
+        if (max >= Short.MIN_VALUE && min <= Short.MIN_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((long) Short.MIN_VALUE)));
-        if (max > Byte.MAX_VALUE && min < Byte.MAX_VALUE)
+        if (max >= Byte.MAX_VALUE && min <= Byte.MAX_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((long) Byte.MAX_VALUE)));
-        if (max > Byte.MIN_VALUE && min < Byte.MIN_VALUE)
+        if (max >= Byte.MIN_VALUE && min <= Byte.MIN_VALUE)
             gens.add(Pair.of(1,
                              Gen.cons((long) Byte.MIN_VALUE)));
-        if (max > 0 && min < 0)
+        if (max >= 0 && min <= 0)
             gens.add(Pair.of(1,
                              Gen.cons(0L)));